Rename reference/nonzero_wrapper.hpp to the CamelCase NonzeroWrapper.hpp
